+++ lock4j-core/src/main/java/com/baomidou/lock/annotation/Lock4j.java
@@ -55,10 +55,10 @@ public @interface Lock4j {
     /**
      * @return 过期时间 单位：毫秒
      * <pre>
-     *     过期时间一定是要长于业务的执行时间. 未设置则为默认时间3秒 默认值：{@link Lock4jProperties#expire}
+     *     过期时间一定是要长于业务的执行时间. 未设置则为默认时间30秒 默认值：{@link Lock4jProperties#expire}
      * </pre>
      */
-    long expire() default 0;
+    long expire() default -1;
 
     /**
      * @return 获取锁超时时间 单位：毫秒
@@ -67,5 +67,12 @@ public @interface Lock4j {
      * </pre>
      */
     long acquireTimeout() default -1;
+
+    /**
+     * 业务方法执行完后（方法内抛异常也算执行完）自动释放锁，如果为false，锁将不会自动释放直至到达过期时间才释放 {@link com.baomidou.lock.annotation.Lock4j#expire()}
+     *
+     * @return 是否自动释放锁
+     */
+    boolean autoRelease() default true;
 
 }