+++ lock4j-core/src/main/java/com/baomidou/lock/executor/LockExecutor.java
@@ -25,6 +25,13 @@ package com.baomidou.lock.executor;
 public interface LockExecutor<T> {
 
     /**
+     * 续期，目前只有redisson支持，切expire参数为-1才会续期
+     * @return
+     */
+    default boolean renewal() {
+        return false;
+    }
+    /**
      * 加锁
      *
      * @param lockKey        锁标识